+++ src/event/event_ctx.h
@@ -11,7 +11,7 @@ struct event_ctx;
 typedef struct event_ctx event_ctx_t;
 
 
-typedef int (*wait_func_t)(event_ctx_t*, struct event*, size_t);
+typedef int (*wait_func_t)(event_ctx_t*, struct event*, int);
 typedef void (*add_func_t)(event_ctx_t*, struct event);
 
 
@@ -21,7 +21,7 @@ struct event_ctx {
 };
 
 
-int event_ctx_wait(event_ctx_t*, struct event*, size_t);
+int event_ctx_wait(event_ctx_t*, struct event*, int);
 
 
 void event_ctx_add(event_ctx_t*, struct event);
